// src/utils/enums/todoActionTypes.ts
export enum TodoActionTypes {
  FETCH_TODOS = 'FETCH_TODOS',
  FETCH_TODOS_SUCCESS = 'FETCH_TODOS_SUCCESS',
  FETCH_TODOS_ERROR = 'FETCH_TODOS_ERROR',
  DELETE_TODO = 'DELETE_TODO',
  DELETE_TODO_SUCCESS = 'DELETE_TODO_SUCCESS',
  DELETE_TODO_ERROR = 'DELETE_TODO_ERROR',
  CREATE_TODO = 'CREATE_TODO',
  CREATE_TODO_SUCCESS = 'CREATE_TODO_SUCCESS',
  CREATE_TODO_ERROR = 'CREATE_TODO_ERROR',
  EDIT_TODO = 'EDIT_TODO',
  EDIT_TODO_SUCCESS = 'EDIT_TODO_SUCCESS',
  EDIT_TODO_ERROR = 'EDIT_TODO_ERROR',
}